Set parameters only inside table

+++ gnuplot.lua
@@ -7,52 +7,10 @@ gnuplot.type = 'gnuplot'
 
 local FUNCTION, DATA = 'function', 'data'
 
-local function tocsv(fname, tbl)
-   local f = io.open(fname, 'w')
-   for _, row in ipairs(tbl) do
-      for i, val in ipairs(row) do
-         f:write((i>1 and ',' or ''), val)
-      end
-      f:write('\n'); f:flush()
-   end
-   f:close()
-end
-
 function gnuplot:new(o)
    local o = o or {}
    setmetatable(o, self)
    return o
-end
-
-gnuplot.cmd = function (g, cmd)
-   if g.handle then
-      g.handle:write(cmd, '\n')
-      g.handle:flush()
-   end
-end
-
-gnuplot.set_fn = function (g, f, l) g[#g+1] = {f, legend=l, type=FUNCTION} end
-
-gnuplot.set_src = function (g, s, l) g[#g+1] = {s, legend=l, type=DATA} end
-
-gnuplot.set_xrange = function (g, x1, x2, dx) g.xrange = {x1, x2, dx} end
-
-gnuplot.set_yrange = function (g, y1, y2, dy) g.yrange = {y1, y2, dy} end
-
-gnuplot.set_xlabel = function (g, lbl) g.xlabel = lbl end
-
-gnuplot.set_ylabel = function (g, lbl) g.ylabel = lbl end
-
-gnuplot.set_title = function (g, txt) g.title = txt end
-
-gnuplot.set_smooth = function (g, sm) g.smooth = sm end
-
-gnuplot.set_separator = function (g, sep) g.separator = sep end
-
-gnuplot.set_permanent = function (g, b) g.permanent = (b or true) end
-
-gnuplot.close = function (g) 
-   if g.handle then g.handle:close() end 
 end
 
 local function tmptable(t)
@@ -76,8 +34,7 @@ end
 
 gnuplot.plot = function (g)
    -- reopen window
-   if g.handle then g.handle:close() end
-   g.handle = io.popen('gnuplot' .. (g.permanent and ' -p' or ''), 'w')
+   local handle = io.popen('gnuplot' .. (g.permanent and ' -p' or ''), 'w')
    -- settings
    local cmd = {}
    g.xrange = g.xrange or {-10,10}
@@ -93,7 +50,7 @@ gnuplot.plot = function (g)
    local file_list = {}
    for _,f in ipairs(g) do
       local str = ''
-      if not f.type or f.type == FUNCTION then
+      if not f.type then
          if type(f[1]) == 'string' then str = f[1]
 	 elseif type(f[1]) == 'function' then
 	    g.xrange = g.xrange or {-10,10}
@@ -111,35 +68,54 @@ gnuplot.plot = function (g)
 	 end
 	 str = '\"'..str..'\" smooth unique'
       end
-      if f.legend then str = string.format('%s title "%s"', str, f.legend) end
+      if f.title then str = string.format('%s title "%s"', str, f.title) end
       plot[#plot+1] = str
    end
    local plot_fn = 'plot ' .. table.concat(plot, ',')
    cmd[#cmd+1] = plot_fn
-   gnuplot.cmd(g, table.concat(cmd,'\n'))
-   --print(table.concat(cmd, '\n'))
-   os.execute('sleep 1')  -- wait for image window
-   for _, f in ipairs(file_list) do os.remove(f) end
+   -- execute
+   handle:write(table.concat(cmd,'\n'),'\n')
+   handle:flush()
+   -- free resources
+   if #file_list > 0 then
+      os.execute('sleep 1')  -- waiting for image window
+      for _, f in ipairs(file_list) do os.remove(f) end
+   end
+   handle:close()
    
    return getmetatable(g) and g or gnuplot:new(g)
 end
 
+gnuplot.__tostring = function (g) 
+   local res = {'{'}
+   for _, f in ipairs(g) do
+      local ftbl = {}
+      for i = 1, #f do table.insert(ftbl, tostring(f[i])) end
+      if f.title then table.insert(ftbl,"'"..f.title.."'") end
+      res[#res+1] = ' {'..table.concat(ftbl, ', ')..'}'
+   end
+   if g.title then res[#res+1] = " title='"..g.title.."'" end
+   if g.xrange then res[#res+1] = ' xrange={'..table.concat(g.xrange,',')..'}' end
+   if g.yrange then res[#res+1] = ' yrangle={'..table.concat(g.yrange,'.')..'}' end
+   if g.xlabel then res[#res+1] = " xlabel='"..g.xlabel.."'" end
+   if g.ylabel then res[#res+1] = " ylable='"..g.ylabel.."'" end
+   res[#res+1] = ' permanent='..(g.permanent==true and 'true' or 'false')
+   res[#res+1] = '}'
+   return table.concat(res, '\n')
+end
+
 ----------------------------------------
 
---[[
-g = gnuplot:new()
-g:set_xrange(0, 5)
-g:set_xlabel("A")
-g:set_ylabel("B")
-g:set_permanent()
-g:set_fn('sin(x)*x', 'legend')
-g:set_fn('cos(x)', 'other')
-g:set_title("Test")
+g = gnuplot.plot
+{
+   {'sin(x)', title ='sinus x'},
+   {'cos(x)', title ='cosinus x'},
+   title = "Test",
+   xrange = {0,5},
+   yrange = {-1,1},
+   xlabel = "X",
+   ylabel = "Y",
+   permanent=true,
+}
 
-g:plot()
-]]
-
-g = gnuplot:new()
-g:set_permanent()
-g:set_fn(math.sin, "sinus")
-g:plot()
+print(g)
